requirements.txt
asgiref==3.6.0
attrs==22.2.0
Django==4.1.5
django-phonenumber-field==7.0.2
djangorestframework==3.14.0
exceptiongroup==1.1.0
iniconfig==2.0.0
packaging==23.0
phonenumbers==8.13.4
pluggy==1.0.0
pytest==7.2.1
pytest-django==4.5.2
python-dotenv==0.21.1
pytz==2022.7.1
sqlparse==0.4.3
tomli==2.0.1
